refactor(libsy): assemble the escalation router as a fall-through cascade
