+++ crates/libsy/src/algorithms/util/escalation.rs
@@ -1,0 +1,605 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+//! Trajectory-judge components for the escalation router — the judge itself, the verdict
+//! policy, the confirmation streak, and the transcript condenser they read.
+//!
+//! [`ConfirmedJudge`] is the whole cascade-facing surface: it calls the judge, applies the
+//! confirmation policy, and scores the capable tier only once a session has confirmed.
+//! Everything else here is private to that. The assembled algorithm lives in
+//! [`crate::algorithms::escalation`].
+
+use std::collections::HashMap;
+
+use async_trait::async_trait;
+use parking_lot::Mutex;
+use serde::Deserialize;
+use switchyard_protocol::{ContentBlock, LlmRequest, Message, OutputParams, Role};
+
+use super::{load_judge_config, Judge, JudgeClassifier, JudgeConfig, JudgePolicy};
+use crate::{Classification, Classifier, Driver, LlmTarget, Request, Result, Score, State};
+
+const PROMPT_TEMPLATE: &str = include_str!("../../prompts/escalation/prompt.md");
+const SCHEMA_TEMPLATE: &str = include_str!("../../prompts/escalation/schema.json");
+
+/// Separator marking where [`truncate_middle`] dropped a message's interior.
+const TRIM_MARKER: &str = " ...[trimmed] ";
+
+/// Suffix marking a transcript cut off by [`MAX_REQUEST_CHARS`].
+const TRUNCATION_SUFFIX: &str = "...<truncated>";
+
+/// Per-message cap for system and developer anchors. Coding-agent harnesses inject very large
+/// boilerplate system prompts carrying no trajectory signal; uncapped they crowd out the window.
+const SYSTEM_CHARS: usize = 1_000;
+
+/// Cap for the first user message — the task statement the judge needs to detect drift, so it
+/// gets the most generous anchor budget.
+const FIRST_USER_CHARS: usize = 2_000;
+
+/// Backstop on the whole assembled transcript, for a pathological single message. The window
+/// caps below normally bind first; when this does bind, the oldest window lines drop.
+const MAX_REQUEST_CHARS: usize = 18_000;
+
+/// Completion budget for one judge reply, covering any reasoning the judge model emits
+/// alongside the verdict.
+///
+/// A runaway guard, not a budget: output tokens cost what they generate, so a generous cap is
+/// nearly free, while a tight one truncates mid-reasoning into unparseable JSON and fails the
+/// judge open on every call. Sized well above the verdict itself — the benchmarked judge's
+/// `reason` ran ~40 tokens at the median and ~136 at its longest — leaving the remainder as
+/// reasoning headroom.
+const JUDGE_MAX_OUTPUT_TOKENS: u64 = 4_096;
+
+/// Ceiling on tracked confirmation streaks, mirroring `AffinityRouter`'s assignment cap.
+const MAX_STREAKS: usize = 4_096;
+
+/// The tuning surface for the trajectory judge.
+///
+/// Deliberately small: these are the three settings an audit of the Python router's twenty-odd
+/// knobs found to actually change routing outcomes. Everything else it exposed is either a
+/// fixed invariant (see the `*_CHARS` constants above) or operational plumbing.
+///
+/// Defaults are the benchmarked configuration; omitted fields take those values.
+#[derive(Clone, Debug, Deserialize)]
+#[serde(default)]
+pub struct EscalationJudgeSettings {
+    /// Consecutive escalate verdicts required before the latch fires.
+    ///
+    /// `1` latches on the first verdict. Higher values filter one-shot eager verdicts — a
+    /// single failed command misread as a pattern — while keeping recall on real trouble,
+    /// which by definition persists across turns. On the benchmarked workload `2` suppressed
+    /// roughly two thirds of escalate verdicts, so this is the router's main cost dial.
+    ///
+    /// The streak is keyed on the session id and held per process. A request without a
+    /// session id cannot accumulate one, so at `2` or higher it can never escalate.
+    pub confirmations: u32,
+    /// Trailing messages shown on top of the anchors. Loop detection needs to see the
+    /// repeats, so a cycle longer than this window is invisible to the judge.
+    pub recent_turn_window: usize,
+    /// Per-message cap inside the trailing window. Error signatures and command shapes
+    /// survive this easily; full file dumps do not need to.
+    pub window_message_chars: usize,
+}
+
+impl EscalationJudgeSettings {
+    /// Rejects settings that would leave the judge with nothing useful to read.
+    fn validate(&self) -> Result<()> {
+        let reject = |message: String| Err(crate::LibsyError::AlgorithmError { message });
+        if self.confirmations == 0 {
+            return reject("confirmations must be at least 1".to_string());
+        }
+        if self.recent_turn_window == 0 {
+            return reject("recent_turn_window must be at least 1".to_string());
+        }
+        if self.window_message_chars < 50 {
+            return reject(format!(
+                "window_message_chars must be at least 50, got {}",
+                self.window_message_chars
+            ));
+        }
+        Ok(())
+    }
+}
+
+impl Default for EscalationJudgeSettings {
+    fn default() -> Self {
+        Self {
+            confirmations: 2,
+            recent_turn_window: 28,
+            window_message_chars: 500,
+        }
+    }
+}
+
+#[derive(Deserialize)]
+#[serde(deny_unknown_fields)]
+struct EscalationVerdict {
+    escalate: bool,
+    #[allow(dead_code)]
+    reason: String,
+}
+
+/// Builds the judge request from the conversation so far.
+struct EscalationJudge {
+    config: JudgeConfig,
+    settings: EscalationJudgeSettings,
+}
+
+impl Judge for EscalationJudge {
+    type Verdict = EscalationVerdict;
+
+    /// Renders the rubric as the system message and the condensed trajectory as the user
+    /// message. `state` is unused: the judge reads only the live request.
+    fn build_request(&self, _state: &State, request: &Request) -> Request {
+        let messages = &request.llm_request.messages;
+        let summary = summarize_for_judge(messages, conversation_turn(request), &self.settings);
+        Request {
+            llm_request: LlmRequest {
+                model: request.llm_request.model.clone(),
+                messages: vec![
+                    Message::text(Role::System, self.config.system_prompt.clone()),
+                    Message::text(Role::User, summary),
+                ],
+                output: OutputParams {
+                    response_format: self.config.response_schema.clone(),
+                    max_output_tokens: Some(JUDGE_MAX_OUTPUT_TOKENS),
+                },
+                ..LlmRequest::default()
+            },
+            raw_request: None,
+            metadata: request.metadata.clone(),
+        }
+    }
+}
+
+/// Maps the judge's verdict to a routing classification, keeping "the judge declined" and
+/// "the judge was unavailable" distinguishable.
+///
+/// Both route to the efficient tier, but only a decline is evidence about the run, so only a
+/// decline clears a confirmation streak. [`Classification::Ambiguous`] carries the unavailable
+/// case: it argmaxes to nothing, exactly like an empty score set, but the router can tell them
+/// apart.
+struct EscalationPolicy {
+    capable: String,
+}
+
+impl JudgePolicy for EscalationPolicy {
+    type Verdict = EscalationVerdict;
+
+    fn to_classification(&self, verdict: Option<&EscalationVerdict>) -> Classification {
+        match verdict {
+            Some(verdict) if verdict.escalate => Classification::Scores(vec![Score {
+                target: self.capable.clone(),
+                confidence: 1.0,
+            }]),
+            Some(_) => Classification::Scores(Vec::new()),
+            None => Classification::Ambiguous(Vec::new()),
+        }
+    }
+}
+
+/// Escalates once a session's escalate verdicts have confirmed.
+///
+/// Abstains — deferring to whatever follows it in the cascade — on a decline, on an escalate
+/// verdict that has not yet confirmed, and on an unavailable judge. Only a decline is evidence
+/// about the run, so only a decline clears the streak.
+pub(crate) struct ConfirmedJudge {
+    classifier: JudgeClassifier<EscalationJudge, EscalationPolicy>,
+    /// Consecutive escalate verdicts required to latch.
+    confirmations: u32,
+    /// Consecutive escalate verdicts per session, cleared by any decline. Only consulted when
+    /// `confirmations > 1`; a single-verdict latch needs no bookkeeping.
+    streaks: Mutex<HashMap<String, u32>>,
+}
+
+impl ConfirmedJudge {
+    /// Builds the judge over `judge_target`, scoring `capable` when it escalates.
+    ///
+    /// Loads the packaged prompt and schema, so an unusable asset or an unusable `settings`
+    /// value fails here rather than on the first request.
+    pub(crate) fn new(
+        judge_target: LlmTarget,
+        capable: String,
+        settings: EscalationJudgeSettings,
+    ) -> Result<Self> {
+        settings.validate()?;
+        let config = load_judge_config(PROMPT_TEMPLATE, SCHEMA_TEMPLATE)?;
+        let confirmations = settings.confirmations;
+        Ok(Self {
+            classifier: JudgeClassifier::new(
+                EscalationJudge { config, settings },
+                judge_target,
+                EscalationPolicy { capable },
+            ),
+            confirmations,
+            streaks: Mutex::new(HashMap::new()),
+        })
+    }
+
+    /// Records one escalate verdict and reports whether the session has now confirmed.
+    ///
+    /// A session id is required to accumulate across turns, so without one nothing can confirm
+    /// beyond a single-verdict latch.
+    fn confirm(&self, request: &Request) -> bool {
+        if self.confirmations <= 1 {
+            return true;
+        }
+        let Some(session) = session_id(request) else {
+            return false;
+        };
+        let mut streaks = self.streaks.lock();
+        if streaks.len() >= MAX_STREAKS && !streaks.contains_key(session) {
+            if let Some(evicted) = streaks.keys().next().cloned() {
+                streaks.remove(&evicted);
+            }
+        }
+        let streak = streaks.entry(session.to_string()).or_insert(0);
+        *streak += 1;
+        *streak >= self.confirmations
+    }
+
+    /// Clears a session's streak after a decline. Strict-consecutive: any decline resets.
+    fn clear_streak(&self, request: &Request) {
+        if self.confirmations <= 1 {
+            return;
+        }
+        if let Some(session) = session_id(request) {
+            self.streaks.lock().remove(session);
+        }
+    }
+}
+
+#[async_trait]
+impl Classifier<State> for ConfirmedJudge {
+    async fn score(
+        &self,
+        state: &mut State,
+        request: &mut Request,
+        driver: Option<&Driver>,
+    ) -> Result<Classification> {
+        let classification = self.classifier.score(state, request, driver).await?;
+        let escalate = match classification {
+            Classification::Scores(ref scores) if !scores.is_empty() => self.confirm(request),
+            Classification::Scores(_) => {
+                self.clear_streak(request);
+                false
+            }
+            // Judge unavailable: fail open to efficient without disturbing the streak.
+            Classification::Ambiguous(_) => false,
+        };
+        if escalate {
+            // The policy already named the capable target; pass its verdict through.
+            Ok(classification)
+        } else {
+            Ok(Classification::Scores(Vec::new()))
+        }
+    }
+}
+
+/// The caller-supplied conversation id, when present and non-empty.
+fn session_id(request: &Request) -> Option<&str> {
+    request
+        .metadata
+        .as_ref()
+        .and_then(|metadata| metadata.session_id.as_deref())
+        .filter(|id| !id.is_empty())
+}
+
+/// The 1-indexed model invocation this request represents: one past each assistant reply.
+///
+/// libsy receives messages already normalized by `switchyard-protocol`, so unlike the
+/// wire-format-aware Python equivalent this needs no per-format branching.
+fn conversation_turn(request: &Request) -> usize {
+    request
+        .llm_request
+        .messages
+        .iter()
+        .filter(|message| message.role == Role::Assistant)
+        .count()
+        + 1
+}
+
+/// Flattens a message to plain text, tool calls and tool results included.
+///
+/// [`Message::text_content`] is deliberately not used here: it keeps only text and refusal
+/// blocks, which would erase exactly the repeated-command signal the judge's loop detection
+/// relies on.
+fn message_text(message: &Message) -> String {
+    let mut parts = Vec::new();
+    collect_text(&message.content, &mut parts);
+    parts.join(" ")
+}
+
+/// Appends the judge-relevant text of each block, descending into tool results.
+fn collect_text(content: &[ContentBlock], parts: &mut Vec<String>) {
+    for block in content {
+        match block {
+            ContentBlock::Text { text } | ContentBlock::Refusal { text } => {
+                parts.push(text.clone());
+            }
+            ContentBlock::ToolCall(call) => {
+                parts.push(format!("tool_call {}({})", call.name, call.arguments));
+            }
+            ContentBlock::ToolResult(result) => collect_text(&result.content, parts),
+            _ => {}
+        }
+    }
+}
+
+/// Keeps the head and tail of `text` within `limit` characters.
+///
+/// The head gets two thirds of the surviving budget: for a trajectory judge the command or
+/// error signature that opens a message carries more signal than its trailing output.
+fn truncate_middle(text: &str, limit: usize) -> String {
+    let chars: Vec<char> = text.chars().collect();
+    if chars.len() <= limit {
+        return text.to_string();
+    }
+    let keep = limit
+        .saturating_sub(TRIM_MARKER.chars().count())
+        .max(20)
+        .min(chars.len());
+    let head = keep * 2 / 3;
+    let tail = keep - head;
+    let mut out: String = chars[..head].iter().collect();
+    out.push_str(TRIM_MARKER);
+    out.extend(chars[chars.len() - tail..].iter());
+    out
+}
+
+/// Renders a compact role-labelled transcript for the judge.
+///
+/// The framing anchors — system/developer messages and the first user message, where agent
+/// harnesses put the task statement — are kept unconditionally and capped individually. The
+/// trailing window carries recent activity. A coverage header states how much history is not
+/// shown, so the judge can reason about pace rather than assuming it sees everything.
+///
+/// When the assembled text still exceeds `max_request_chars`, the oldest window lines go
+/// first: for a trajectory judge the newest evidence is strictly the most valuable.
+fn summarize_for_judge(
+    messages: &[Message],
+    turn: usize,
+    settings: &EscalationJudgeSettings,
+) -> String {
+    let mut anchors: Vec<String> = Vec::new();
+    let mut window: Vec<String> = Vec::new();
+    let mut first_user_seen = false;
+
+    for message in messages {
+        let text = message_text(message);
+        match message.role {
+            Role::System | Role::Developer => anchors.push(format!(
+                "[{}] {}",
+                role_label(message.role),
+                truncate_middle(&text, SYSTEM_CHARS)
+            )),
+            Role::User if !first_user_seen => {
+                first_user_seen = true;
+                anchors.push(format!(
+                    "[user (task)] {}",
+                    truncate_middle(&text, FIRST_USER_CHARS)
+                ));
+            }
+            role => window.push(format!(
+                "[{}] {}",
+                role_label(role),
+                truncate_middle(&text, settings.window_message_chars)
+            )),
+        }
+    }
+
+    if window.len() > settings.recent_turn_window {
+        window.drain(..window.len() - settings.recent_turn_window);
+    }
+
+    let header = format!(
+        "Conversation turn {turn}; showing the last {} of {} messages after the task framing.",
+        window.len(),
+        messages.len(),
+    );
+    let assemble = |window: &[String]| {
+        std::iter::once(header.as_str())
+            .chain(anchors.iter().map(String::as_str))
+            .chain(window.iter().map(String::as_str))
+            .collect::<Vec<_>>()
+            .join("\n")
+    };
+
+    let mut text = assemble(&window);
+    while text.chars().count() > MAX_REQUEST_CHARS && !window.is_empty() {
+        window.remove(0);
+        text = assemble(&window);
+    }
+    if text.chars().count() > MAX_REQUEST_CHARS {
+        let keep = MAX_REQUEST_CHARS.saturating_sub(TRUNCATION_SUFFIX.chars().count() + 1);
+        text = text.chars().take(keep).collect::<String>() + TRUNCATION_SUFFIX;
+    }
+    text
+}
+
+/// The transcript label for a role.
+fn role_label(role: Role) -> &'static str {
+    match role {
+        Role::System => "system",
+        Role::Developer => "developer",
+        Role::User => "user",
+        Role::Assistant => "assistant",
+        Role::Tool => "tool",
+    }
+}
+
+/// A request whose conversation sits at `turn`: `turn - 1` prior assistant replies, each
+/// answered by a further user message.
+///
+/// Shared with the assembled router's tests, which drive the same conversation shape.
+#[cfg(test)]
+pub(crate) fn request_at_turn(session_id: Option<&str>, turn: usize) -> Request {
+    use switchyard_protocol::Metadata;
+
+    let mut messages = vec![Message::text(Role::User, "What is 2+2?")];
+    for attempt in 1..turn {
+        messages.push(Message::text(Role::Assistant, format!("attempt {attempt}")));
+        messages.push(Message::text(Role::User, format!("still wrong {attempt}")));
+    }
+    Request {
+        llm_request: LlmRequest {
+            model: Some("auto".to_string()),
+            messages,
+            ..LlmRequest::default()
+        },
+        raw_request: None,
+        metadata: session_id.map(|id| Metadata {
+            session_id: Some(id.to_string()),
+            ..Metadata::default()
+        }),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use serde_json::json;
+    use switchyard_protocol::{ContentBlock, Message, Role, ToolCall, ToolResult};
+
+    use super::*;
+
+    #[test]
+    fn judge_request_is_rubric_plus_summary_under_a_completion_cap() -> Result<()> {
+        let judge = EscalationJudge {
+            config: load_judge_config(PROMPT_TEMPLATE, SCHEMA_TEMPLATE)?,
+            settings: EscalationJudgeSettings::default(),
+        };
+
+        let built = judge.build_request(&State::default(), &request_at_turn(None, 4));
+
+        // Two messages: the rubric as system, the condensed trajectory as user.
+        assert_eq!(built.llm_request.messages.len(), 2);
+        assert_eq!(built.llm_request.messages[0].role, Role::System);
+        assert_eq!(built.llm_request.messages[1].role, Role::User);
+        assert!(built.llm_request.messages[1]
+            .text_content("")
+            .is_some_and(|text| text.contains("Conversation turn 4")));
+        // Bounded output, so a reasoning judge cannot run away mid-verdict.
+        assert_eq!(
+            built.llm_request.output.max_output_tokens,
+            Some(JUDGE_MAX_OUTPUT_TOKENS)
+        );
+        assert!(built.llm_request.output.response_format.is_some());
+        Ok(())
+    }
+
+    #[test]
+    fn conversation_turn_counts_assistant_replies() {
+        assert_eq!(conversation_turn(&request_at_turn(None, 1)), 1);
+        assert_eq!(conversation_turn(&request_at_turn(None, 5)), 5);
+    }
+
+    #[test]
+    fn message_text_keeps_tool_calls_and_results() {
+        let call = Message {
+            role: Role::Assistant,
+            content: vec![
+                ContentBlock::Text {
+                    text: "running it".to_string(),
+                },
+                ContentBlock::ToolCall(ToolCall {
+                    id: "call-1".to_string(),
+                    name: "bash".to_string(),
+                    arguments: json!({"cmd": "ls"}),
+                }),
+            ],
+        };
+        let text = message_text(&call);
+        assert!(text.contains("running it"), "{text}");
+        assert!(text.contains(r#"tool_call bash({"cmd":"ls"})"#), "{text}");
+
+        let result = Message {
+            role: Role::Tool,
+            content: vec![ContentBlock::ToolResult(ToolResult {
+                tool_call_id: "call-1".to_string(),
+                content: vec![ContentBlock::Text {
+                    text: "no such file".to_string(),
+                }],
+                is_error: Some(true),
+            })],
+        };
+        assert_eq!(message_text(&result), "no such file");
+    }
+
+    #[test]
+    fn truncate_middle_keeps_head_and_tail() {
+        let text = "a".repeat(40) + &"z".repeat(40);
+        let trimmed = truncate_middle(&text, 50);
+        assert!(trimmed.chars().count() <= 50, "{trimmed}");
+        assert!(trimmed.starts_with('a'));
+        assert!(trimmed.ends_with('z'));
+        assert!(trimmed.contains("[trimmed]"));
+
+        // Under the limit the text is returned untouched.
+        assert_eq!(truncate_middle("short", 50), "short");
+    }
+
+    #[test]
+    fn summary_keeps_anchors_and_the_recent_window() {
+        let mut messages = vec![
+            Message::text(Role::System, "you are a coding agent"),
+            Message::text(Role::User, "fix the failing test"),
+        ];
+        for i in 0..10 {
+            messages.push(Message::text(Role::Assistant, format!("step {i}")));
+        }
+        let settings = EscalationJudgeSettings {
+            recent_turn_window: 3,
+            ..EscalationJudgeSettings::default()
+        };
+
+        let summary = summarize_for_judge(&messages, 11, &settings);
+
+        assert!(
+            summary.contains("[system] you are a coding agent"),
+            "{summary}"
+        );
+        assert!(
+            summary.contains("[user (task)] fix the failing test"),
+            "{summary}"
+        );
+        assert!(summary.contains("Conversation turn 11; showing the last 3 of 12 messages"));
+        // Only the newest window entries survive.
+        assert!(summary.contains("step 9"), "{summary}");
+        assert!(summary.contains("step 7"), "{summary}");
+        assert!(!summary.contains("step 6"), "{summary}");
+    }
+
+    #[test]
+    fn summary_drops_oldest_window_lines_under_the_char_cap() {
+        // MAX_REQUEST_CHARS is a backstop, not a dial: at default settings the window caps
+        // bind first (28 x 500 plus anchors sits under it), so reaching it takes an unusually
+        // wide per-message cap. That is the point — it only fires on pathological input.
+        let mut messages = vec![
+            Message::text(Role::System, "framing"),
+            Message::text(Role::User, "task"),
+        ];
+        for i in 0..20 {
+            messages.push(Message::text(
+                Role::Assistant,
+                format!("{i} {}", "x".repeat(2_000)),
+            ));
+        }
+        let settings = EscalationJudgeSettings {
+            window_message_chars: 2_000,
+            ..EscalationJudgeSettings::default()
+        };
+
+        let summary = summarize_for_judge(&messages, 21, &settings);
+
+        assert!(
+            summary.chars().count() <= MAX_REQUEST_CHARS,
+            "{}",
+            summary.chars().count()
+        );
+        // Anchors are never dropped, and the newest activity outlives the oldest.
+        assert!(summary.contains("[system] framing"), "{summary}");
+        assert!(summary.contains("[user (task)] task"), "{summary}");
+        assert!(summary.contains("19 xxx"), "{summary}");
+        assert!(!summary.contains("0 xxx"), "{summary}");
+    }
+}
